Measuring text size by drawable modifier.

+++ SharpCompose.Base/ComposesApi/Providers/LocalTextStyleProvider.cs
@@ -6,6 +6,6 @@ namespace SharpCompose.Base.ComposesApi.Providers;
 public class TextStyle
 {
     public Color? Color { get; init; } = null;
-    public int FontSize { get; init; } = 14;
+    public double FontSize { get; init; } = 14;
     public Font Font { get; init; } = new Font("Helvetica", FontWeight.Regular);
 }
